CameraTrapSimulator/requirements.txt
opencv-python==4.4.0.46
imutils==0.5.3
tensorflow==2.4.0
numpy==1.19.3
pandas==1.1.4
azure-common==1.1.26
azure-core==1.10.0
azure-iot-device==2.5.0